refactor(backend): derive LLM types from Zod schemas (single source of truth)
Replace interface + duplicate schema with z.infer<typeof schema> for
all persisted LLM types. Schemas live next to their source types in
llm-api/ and llm-session/, imported by agent/types.ts.
